+++ src/Examples/BookstoreExample/BookstoreUi/AuthorCrudBuilder.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using TheXDS.Triton.CrudGen;
+﻿using TheXDS.Triton.CrudGen;
 using TheXDS.Triton.CrudGen.Base;
 using TheXDS.Triton.Examples.BookstoreExample.Models;
 
@@ -23,20 +22,6 @@ namespace BookstoreUi
             editor.Property(p => p.Books)
                 .Label("Libros del autor")
                 .CreateAndSelect();
-        }
-    }
-
-    public class BookCrudBuilder : CrudBuilder<Book>
-    {
-        protected override void Describe(ICrudDescriptionBuilder<Book> editor, ICrudDescriptionBuilder<Book> viewer)
-        {
-            editor.Property(p => p.Name)
-                .Label("Nombre")
-                .Big();
-
-            editor.Property(p => p.Edition)
-                .Label("Edición")
-                .Range(1, 10);
         }
     }
 }
